// src/api/back/models/DroppedItemDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    MarketItemDto,
    MarketItemDtoFromJSON,
    MarketItemDtoFromJSONTyped,
    MarketItemDtoToJSON,
} from './';

/**
 * 
 * @export
 * @interface DroppedItemDto
 */
export interface DroppedItemDto {
    /**
     * 
     * @type {string}
     * @memberof DroppedItemDto
     */
    assetId: string;
    /**
     * 
     * @type {number}
     * @memberof DroppedItemDto
     */
    matchId: number;
    /**
     * 
     * @type {string}
     * @memberof DroppedItemDto
     */
    droppedAt: string;
    /**
     * 
     * @type {string}
     * @memberof DroppedItemDto
     */
    activeTradeId: string;
    /**
     * 
     * @type {string}
     * @memberof DroppedItemDto
     */
    expires: string;
    /**
     * 
     * @type {MarketItemDto}
     * @memberof DroppedItemDto
     */
    item: MarketItemDto;
}

export function DroppedItemDtoFromJSON(json: any): DroppedItemDto {
    return DroppedItemDtoFromJSONTyped(json, false);
}

export function DroppedItemDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): DroppedItemDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'assetId': json['assetId'],
        'matchId': json['matchId'],
        'droppedAt': json['droppedAt'],
        'activeTradeId': json['activeTradeId'],
        'expires': json['expires'],
        'item': MarketItemDtoFromJSON(json['item']),
    };
}

export function DroppedItemDtoToJSON(value?: DroppedItemDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'assetId': value.assetId,
        'matchId': value.matchId,
        'droppedAt': value.droppedAt,
        'activeTradeId': value.activeTradeId,
        'expires': value.expires,
        'item': MarketItemDtoToJSON(value.item),
    };
}